centralization of logic to meet the MVC pattern
