+++ server/todo-api/controllers/router.ts
@@ -1,6 +1,3 @@
 import express from 'express';
 import controller from './controller';
-import validateBodyRequest from '../middlewares/validateBodyRequest';
-export default express
-  .Router()
-  .post('/',validateBodyRequest,  controller.create)
+export default express.Router().post('/', controller.create);
